+++ packages/react-auth0-base/src/components/index.ts
@@ -1,1 +1,1 @@
-export * from './walletconnect';
+// export * from './walletconnect';
