+++ packages/react-auth0-base/src/components/walletconnect/useWalletConnectOngoingRequests.ts
@@ -1,0 +1,10 @@
+import React, { useContext } from 'react';
+import { OngoingRequestQueueContext } from './OngoingRequestQueueProvider';
+
+export const useWalletConnectOngoingRequests = () => {
+  const electionQueueContext = useContext(OngoingRequestQueueContext);
+  if (!electionQueueContext) {
+    throw new Error('electionQueueContext has to be used within <WalletConnectContext.Provider>');
+  }
+  return electionQueueContext;
+}
