+++ packages/react-auth0-base/src/index.ts
@@ -1,5 +1,9 @@
 // don't export the base
 // export * from './auth0-sdk-provider-base';
-export * from './sdkClientContext';
+import KrayonSdkClientContext from './sdkClientContext';
+
+export { KrayonSdkClientContext };
+
 export * from './use-sdk-hooks';
 export * from './components';
+export * from './auth0-sdk-provider-base';
