+++ packages/react-native-auth0-sdk/package.json
@@ -6,8 +6,8 @@
   "license": "MIT",
   "homepage": "https://krayondigital.com",
   "peerDependencies": {
-    "@types/react": "^18",
-    "react": "^18",
+    "@types/react": "^18.2.14",
+    "react": "^18.2.14",
     "react-native": "0.72.6",
     "react-native-auth0": "^3.0.2"
   }
