+++ packages/react-native-auth0-sdk/src/index.ts
@@ -1,2 +1,3 @@
-export * from '../../react-auth0-base/src/';
-export * from './react-native-auth0-sdk-provider';
+export * from '../../react-auth0-base/src/index';
+// TODO: add this file
+// export * from '../../src/shared/react-native-auth0-sdk-provider';
