+++ packages/react-native-auth0-sdk/tsconfig.lib.json
@@ -18,5 +18,5 @@
     "**/*.spec.jsx",
     "**/*.test.jsx"
   ],
-  "include": ["**/*.js", "**/*.jsx", "**/*.ts", "**/*.tsx"]
+  "include": ["**/*.js", "**/*.jsx", "**/*.ts", "**/*.tsx", "../../../src/shared/react-native-auth0-sdk-provider.tsx"]
 }
